Fixed a few glitches with the processor. Still randomly teleports and doesn't give the processed good. Also need to fix the broken armor stands
