+++ src/main/java/droidco/west3/ironsight/IronSight.java
@@ -22,6 +22,7 @@ import droidco.west3.ironsight.Horse.HorseEvents;
 import droidco.west3.ironsight.Items.MasterList.MasterListCmd;
 import droidco.west3.ironsight.Items.MasterList.MasterListEvents;
 import droidco.west3.ironsight.NPC.NPCEvents;
+import droidco.west3.ironsight.Processors.ProcessorEvents;
 import droidco.west3.ironsight.Tracker.TrackerEvents;
 import org.bukkit.Bukkit;
 import org.bukkit.command.ConsoleCommandSender;
@@ -85,6 +86,7 @@ public final class IronSight extends JavaPlugin {
         getServer().getPluginManager().registerEvents(new NPCEvents(this), this);
         getServer().getPluginManager().registerEvents(new HorseEvents(), this);
         getServer().getPluginManager().registerEvents(new FishingEvents(), this);
+        getServer().getPluginManager().registerEvents(new ProcessorEvents(this), this);
         getServer().getPluginManager().registerEvents(new VaultEvents(), this);
     }
     public void loadAllCommands() {
